requirements.txt
Django==2.1.3
django-developer-panel==0.1.2
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
psycopg2==2.7.6.1
PyJWT==1.6.4
pytz==2018.7
